admin_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt,argon2]
python-jose[cryptography]
email-validator
argon2-cffi
fastapi-mail
requests
lxml
gunicorn
python-multipart